Feat: update popup component
