+++ src/pages/Newtab/components/Popup.jsx
@@ -1,17 +1,14 @@
 import React, { useState, useEffect } from 'react';
-import { Textarea, Input, Button, Badge } from '@chakra-ui/react';
+import { Textarea, Button } from '@chakra-ui/react';
+import TagInput from './TagInput';
+
 /** @jsx jsx */
 import { css, jsx } from '@emotion/react';
 import CloseBtn from '../../../assets/img/closeBtn.svg';
+import MemoComplete from '../../../assets/img/memoComplete.svg';
 
 const textareaWrapper = css`
   padding: 3rem 1rem 0 1rem;
-`;
-
-const inputWrapper = css`
-  display: flex;
-  justify-content: space-between;
-  padding: 1rem;
 `;
 
 const textareaBox = css`
@@ -31,26 +28,10 @@ const textareaBox = css`
   }
 `;
 
-const inputBox = css`
-  width: 16.875rem;
-  height: 2.5rem;
-  margin-right: 1rem;
-  background: #f7f7f7;
-  color: #3c3a37;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 0.875rem;
-  line-height: 1.188rem;
-  border: none;
-
-  &::placeholder {
-    color: #858585;
-    opacity: 0.4;
-  }
-
-  &.active {
-    padding-left: 6.25rem;
-  }
+const inputWrapper = css`
+  display: flex;
+  justify-content: space-between;
+  padding: 1rem;
 `;
 
 const saveButton = css`
@@ -69,6 +50,30 @@ const saveButton = css`
   }
 `;
 
+const completeWrapper = css`
+  padding: 3.75rem 4.125rem 1.875rem;
+  text-align: center;
+
+  h2 {
+    color: #3c3a37;
+    font-size: 1.5rem;
+    font-weight: 700;
+    font-style: normal;
+    user-select: none;
+  }
+
+  img {
+    margin: 2.688rem 0 1.875rem 0;
+  }
+
+  a {
+    color: #a5aab2;
+    font-size: 0.875rem;
+    font-weight: 400;
+    font-style: normal;
+  }
+`;
+
 const closeButton = css`
   position: absolute;
   top: 1rem;
@@ -80,50 +85,6 @@ const closeButton = css`
   background-repeat: no-repeat;
   cursor: pointer;
 `;
-const keywordTag = css`
-  display: none;
-  height: 1.5rem;
-  padding: 0.125rem 0.5rem;
-  position: absolute;
-  top: 0.5rem;
-  left: 0.813rem;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  border-radius: 2px;
-  cursor: pointer;
-  z-index: 99;
-
-  &.active {
-    display: block;
-  }
-
-  &.red {
-    color: #e64632;
-    background-color: rgba(230, 70, 50, 0.2);
-  }
-`;
-const selectBox = css`
-  width: 15.5rem;
-  position: absolute;
-  background: #fff;
-  border-radius: 4px;
-  margin-top: 0.5rem;
-  box-shadow: 4px 4px 16px rgba(211, 207, 197, 0.7);
-`;
-const selectListItem = css`
-  padding: 0.563rem 0.5rem;
-  cursor: pointer;
-
-  &:hover {
-    background: rgba(60, 58, 55, 0.1);
-  }
-
-  &.active {
-    background: rgba(60, 58, 55, 0.1);
-  }
-`;
 
 const Popup = (props) => {
   const [memo, setMemo] = useState({
@@ -132,23 +93,7 @@ const Popup = (props) => {
     minRows: 6,
     maxRows: 10,
   });
-  const [keyword, setKeyword] = useState('');
-  const [selectKeyword, setSelectKeyword] = useState([]);
-  const [tempArr, setTempArr] = useState([]);
-  const [findFlag, setFindFlag] = useState(false);
-  const [tagFlag, setTagFlag] = useState(false);
-  const [selectPosition, setSelectPosition] = useState(0);
-
-  useEffect(() => {
-    //TODO get folder name list API
-    setTempArr(['디자인', '여행', '맛집', '강아래팀', '강아래']);
-  }, []);
-
-  useEffect(() => {
-    if (keyword !== '') {
-      keywordFind();
-    }
-  }, [keyword]);
+  const [submitFlag, setSubmitFlag] = useState(false);
 
   // memo Textarea size control
   const memoChange = (event) => {
@@ -173,86 +118,20 @@ const Popup = (props) => {
     });
   };
 
-  // keyword input find words
-  const keywordFind = () => {
-    let tempKey = [];
-
-    tempArr.map((item) => {
-      if (keyword === '') {
-        return;
-      } else if (item.includes(keyword)) {
-        tempKey.push(item);
-      }
-      return item;
-    });
-    setSelectKeyword(tempKey);
-  };
-
-  // keyword input keypress event
-  const handleKeyPress = (event) => {
-    if (event.key === 'Enter' && keyword !== '') {
-      if (selectKeyword.length > 0) {
-        setFindFlag(true);
-        setKeyword('');
-      } else {
-        setSelectKeyword([keyword]);
-        setFindFlag(true);
-        setKeyword('');
-      }
-    } else if (event.key === 'Backspace') {
-      if (selectKeyword.length > 0 && tagFlag) {
-        setFindFlag(false);
-        setSelectKeyword([]);
-        setTagFlag(false);
-      }
-      if (!tagFlag) {
-        setTagFlag(true);
-      }
-    } else if (event.key === 'ArrowUp') {
-      // top key press
-      if (selectPosition > 0) {
-        setSelectPosition(selectPosition - 1);
-      }
-    } else if (event.key === 'ArrowDown') {
-      // down key press
-      if (selectKeyword.length > selectPosition + 1) {
-        setSelectPosition(selectPosition + 1);
-      }
+  const writePopupResult = (type) => {
+    if (type === 'submit') {
+      setSubmitFlag(true);
     }
-  };
-
-  // select List click keyword
-  const handleSelectItem = () => {
-    setFindFlag(true);
-    setKeyword('');
-  };
-
-  // focus out keyword input
-  const handleFocusOut = () => {
-    if (keyword !== '') {
-      setSelectKeyword([keyword]);
-      setFindFlag(true);
-      setKeyword('');
+    // chrome-extension popup close
+    if (props.extension && submitFlag) {
+      window.close();
+      return;
+    } else if (!props.extension) {
+      props.setWritePopup({
+        ...props.writePopup,
+        flag: false,
+      });
     }
-  };
-
-  const renderSelectList = () => {
-    let html = [];
-
-    selectKeyword.map((item, idx) => {
-      html.push(
-        <li
-          css={selectListItem}
-          className={selectPosition === idx && 'active'}
-          key={idx}
-          onClick={() => handleSelectItem()}
-        >
-          {item}
-        </li>,
-      );
-    });
-
-    return html;
   };
 
   return (
@@ -263,54 +142,42 @@ const Popup = (props) => {
         minHeight: 280,
         maxHeight: 420,
         position: 'absolute',
-        left: props.mainFlag ? 100 : 0,
-        top: props.mainFlag ? 100 : 0,
+        left: props.writePopup.flag ? props.writePopup.left : 0,
+        top: props.writePopup.flag ? props.writePopup.top : 0,
         boxShadow: '0px 10px 16px rgba(211, 207, 197, 0.7)',
         borderRadius: 2,
         backgroundColor: '#fff',
-        zIndex: 99,
+        zIndex: 9,
       }}
     >
-      <a css={closeButton} onClick={() => props.popupResult('close')}></a>
-      <div css={textareaWrapper}>
-        <Textarea
-          css={textareaBox}
-          placeholder="무엇이 떠오르세요?"
-          focusBorderColor="none"
-          rows={memo.rows}
-          value={memo.value}
-          onChange={memoChange}
-        />
-      </div>
-      <div css={inputWrapper}>
-        <div css={{ position: 'relative' }} className="form-group">
-          <Input
-            css={inputBox}
-            className={findFlag ? 'active' : ''}
-            placeholder={findFlag ? '' : '저장할 폴더의 이름을 입력하세요'}
-            focusBorderColor="black"
-            autoComplete="off"
-            value={keyword}
-            onBlur={() => handleFocusOut()}
-            onKeyDown={(e) => handleKeyPress(e)}
-            onChange={findFlag ? undefined : (e) => setKeyword(e.target.value)}
-          />
-          <Badge css={keywordTag} className={findFlag ? 'active red' : ''}>
-            {/* TODO temp code */}
-            {selectKeyword[selectPosition] + (tagFlag ? 'X' : '')}
-          </Badge>
-          {!findFlag && selectKeyword.length > 0 && (
-            <div css={selectBox}>
-              <ul>{renderSelectList()}</ul>
-            </div>
-          )}
+      <a css={closeButton} onClick={() => writePopupResult('close')}></a>
+      {!submitFlag ? (
+        <>
+          <div css={textareaWrapper}>
+            <Textarea
+              css={textareaBox}
+              placeholder="무엇이 떠오르세요?"
+              focusBorderColor="none"
+              rows={memo.rows}
+              value={memo.value}
+              onChange={memoChange}
+            />
+          </div>
+          <div css={inputWrapper}>
+            <TagInput />
+            <Button css={saveButton} onClick={() => writePopupResult('submit')}>
+              저장
+            </Button>
+          </div>
+        </>
+      ) : (
+        <div css={completeWrapper}>
+          <h2>사각사각에 저장되었어요!</h2>
+          <img src={MemoComplete} alt="Memo Save Complete" />
+          <a>사각사각으로 이동하기</a>
         </div>
-        <Button css={saveButton} onClick={() => props.popupResult('submit')}>
-          저장
-        </Button>
-      </div>
+      )}
     </aside>
   );
 };
-
 export default Popup;
